"Remove AddCourse component and modify ResponsiveAppBar to include a change password modal"
This commit removes the AddCourse component and modifies the ResponsiveAppBar component to include a new password change modal. The password change functionality is implemented in the updated ResponsiveAppBar component. The addDepartment and addSession functions in the admin.js file are also updated to reflect the change in the function parameter name from credentials to departmentDetails and sessionDetails, respectively.